Adding big decimal and other unit tests

+++ src/main/java/com/creditmodule/ing/data/CreateAssetRequest.java
@@ -4,15 +4,17 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Positive;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class CreateAssetRequest {
     @NotBlank(message = "Asset name is required")
     private String assetName;
 
     @Positive(message = "Initial size must be greater than 0")
-    private double initialSize;
+    private BigDecimal initialSize;
 
     @Positive(message = "Initial price must be greater than 0")
-    private double initialPrice;
+    private BigDecimal initialPrice;
 
 }
