+++ src/main/java/com/creditmodule/ing/data/CreateAssetResponse.java
@@ -4,13 +4,15 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class CreateAssetResponse {
     private Long id;
     private String assetName;
-    private double size;
-    private double usableSize;
+    private BigDecimal size;
+    private BigDecimal usableSize;
     private String message;
 }
