+++ src/main/java/com/creditmodule/ing/data/CustomerAssetResponse.java
@@ -4,12 +4,14 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class CustomerAssetResponse {
     private Long assetId;
     private String assetName;
-    private double totalSize;
-    private double usableSize;
+    private BigDecimal totalSize;
+    private BigDecimal usableSize;
 }
