+++ src/main/java/com/creditmodule/ing/data/UserCustomerCreateRequest.java
@@ -2,10 +2,12 @@ package com.creditmodule.ing.data;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class UserCustomerCreateRequest {
     private String surname;
     private String password;
     private String name;
-    private Long creditLimit;
+    private BigDecimal creditLimit;
 }
